# Print relative speed up in benchmarks