--- a/atomic-counter.cabal
+++ b/atomic-counter.cabal
@@ -125,8 +125,8 @@ benchmark bench
     , base >= 4.14
     , primitive
     , stm
-    , tasty
-    , tasty-bench
+    , tasty >= 1.4.2
+    , tasty-bench >= 0.3.2
     , tasty-quickcheck
     , test-utils
   ghc-options: